Move CommandLineFactory into factory package
