+++ telegram-bot/src/main/java/io/github/bivashy/wttj/telegram/bot/command/factory/CommandLineFactory.java
@@ -1,4 +1,4 @@
-package io.github.bivashy.wttj.telegram.bot.command;
+package io.github.bivashy.wttj.telegram.bot.command.factory;
 
 import io.github.bivashy.wttj.api.command.Actor;
 import picocli.CommandLine;
